Corrected bug that prevented table of composite to works properly.

+++ modules/core/include/nt2/core/container/details/block/nrc.hpp
@@ -74,7 +74,7 @@ namespace nt2 { namespace details
   {
     typedef block_nrc < Type
                             , heap_(Allocator)
-                            , boost::mpl::int_<1>
+                            , boost::mpl::size_t<1>
                             , true
                             >                             prior_block;
